# visit-scheduler-api-specification.yaml
openapi: "3.0.2"
info:
  title: Visit Scheduler Service
  version: "1.0"
servers:
  - url: https://visit-scheduler.gov.uk/v1
    variables:
      environment:
        default: test # Test
        enum:
          - dev       #Development
          - test      #Test
          - preprod   #Pre-production
          - Prod      #Production
paths:
  /visit-sessions:
    get:
      security:
        - bearerAuth: []
      summary: Returns visit sessions for the specified prison which are within the bookable time period - whether or not they are full
      tags:
        - visit-sessions
      operationId: getBookableSessions
      parameters:
        - name: prisonId
          in: query
          description: NOMIS prison Identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VisitSession'
        404:
          description: prison not found
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        401:
          description: Authentication required.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        403:
          description: Access denied
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
  /visits/{visitId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Retrieve details of a visit
      tags:
        - visits
      operationId: getVisits
      parameters:
        - name: visitId
          in: path
          description: Identifier for this visit
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Visit'
        404:
          description: Visit identifier not found
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        401:
          description: Authentication required.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        403:
          description: Access denied
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
    put:
      security:
        - bearerAuth: []
      summary: Revise an existing visit
      tags:
        - visits
      operationId: reviseVisit
      parameters:
        - name: visitId
          in: path
          description: identidier for this visit
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Visit'
        400:
          description: Request incorrectly formed.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        401:
          description: Authentication required.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        403:
          description: Access denied
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        404:
          description: Visit identifier not found
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
  /visits:
    get:
      security:
        - bearerAuth: [ ]
      summary: Retrieve visits with optional filtering
      tags:
        - visits
      operationId: getVisitsByPrisoner
      parameters:
        - name: prisonerId
          in: query
          description: NOMIS Prisoner Reference
          required: false
          schema:
            type: string
        - name: prisonId
          in: query
          description: NOMIS prison Reference
          required: false
          schema:
            type: string
        - name: nomisPersonId
          in: query
          description: contact Id (currently NOMIS person id)
          required: false
          schema:
            type: integer
        - name: startTimestamp
          in: query
          description: retrieve visits when visit start is on or after this timestamp (inclusive)
          required: false
          schema:
            type: string
          example: 2021-11-03T08:00:00
        - name: endTimestamp
          in: query
          description: retrieve visits when visit start is on or before this timestamp (inclusive)
          required: false
          schema:
            type: string
          example: 2021-11-03T09:00:00
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Visit'
        400:
          description: Request incorrectly formed
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        401:
          description: Authentication required
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        403:
          description: Access denied
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a new visit
      tags:
        - visits
      operationId: createVisit
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Visit'
        400:
          description: Request incorrectly formed.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        401:
          description: Authentication required.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        403:
          description: Access denied
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NonSuccessResponse'
        404:
          description: Visit Identifier not found
          content:
            application/json:
              schema:
                items:
                $ref: '#/components/schemas/NonSuccessResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    NonSuccessResponse:
      type: object
      properties:
        applicationErrorCode:
          type: string
        applicationErrorDescription:
          type: string
    VisitSession:
      type: object
      properties:
        sessionId:
          type: string
          description: The unique identifier for this session
        prison:
          $ref: '#/components/schemas/prison'
        visitRoomName:
          type: string
          description: The Name of the visit room in which this visit session takes place
        visitType:
          type: string
          description: The type of visits taking place within this session
        openVisitCapacity:
          type: integer
          description: The number of concurrent visits which may take place within this session
        openVisitsAvailableCount:
          type: integer
          description: The count of open visit slots still available for this session
        closedVisitCapacity:
          type: integer
          description: The number of closed visits which may take place within this session
        closedVisitAvailableCount:
          type: integer
          description: The count of closed visit slots still available for this session
        startTimestamp:
          type: string
          format: date-time
          description: The start timestamp for this visit session
        endTimestamp:
          type: string
          format: date-time
          description: The end timestamp for this visit session
    Visit:
      type: object
      properties:
        visitId:
          type: string
          description: Publicised identifier for this visit
        prisonId:
          type: string
          description: The NOMIS identifier for the prison for this visit
        visitType:
          $ref: '#/components/schemas/VisitType'
        prisonerId:
          type: string
          description: Prisoner identifier for the prisoner being visited
        visitRoom:
          type: string
          description: The visit room where this visit is taking place
        visitDate:
          type: string
          format: date
          description: Visit date
        startTime:
          type: string
          format: time
          description: Start time for the visit
        endTime:
          type: string
          format: time
          description: End time for the visit
        visitors:
          description: Set of Visitors
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        reasonableAdjustments:
          type: string
          description: Text notes for reasonable adjustments for this visit
        currentStatus:
          $ref: '#/components/schemas/VisitStatus'
        lastRevisedTimestamp:
          type: string
          description: Timestamp this visit was last revised
          format: date-time
    VisitStatus:
      type: string
      nullable: false
      enum:
        - Reserved
        - Confirmed
        - Cancelled by Prisoner
        - Cancelled by Visitor
        - Cancelled by prison
        - Attended
    VisitType:
      type: string
      nullable: false
      enum:
        - Social
        - Family
        - Official
        - Legal
    Contact:
      type: object
      description: Identifies the contact
      properties:
        nomisPersonId:
          type: string
          description: Unique Identifier for this contact (from NOMIS)
        leadVisitor:
          type: boolean
          description: indicates if this visitor is the lead visitor
    Prisoner:
      type: object
      description: A description
      properties:
        prisonerId:
          type: string
          description: Unique identifier for this prisoner (from NOMIS)
        firstName:
          type: string
          description: Prisoner first name
        middleName:
          type: string
          description: Prisoner middle name
        lastName:
          type: string
          description: Prisoner last name
        dateOfBirth:
          type: string
          description: prisoner date of birth
    prison:
      type: object
      description: prison Description
      properties:
        prisonId:
          type: string
          description: Unique identifier for this prison
        prisonName:
          type: string
          description: prison Name